Problem with Qt applications on 64 bit machines fixed
Dropdown menu elements were not visible on Qt applications. The reason
for that was that the _NET_DESKTOP_VIEWPORT and _NET_DESKTOP_GEOMETRY
values on the root window were wrong. Moreover the the strut values from
applications like gkrellm were also wrong.

Explanation:
Array elements to function XChangeProperty and type XA_CARDINAL must be
of type "glong". Even if we give a bit length of 32 to the call.

+++ src/octopus.h
@@ -49,7 +49,7 @@ typedef struct {
 } rect_t;
 
 typedef struct {
-  gint x, y, w, h;
+  glong x, y, w, h;
 } warea_t;
 
 typedef struct {
